SSCSCI-1845: revert ReadyToListSubmittedHandler changes
