+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/readytolist/ReadyToListAboutToSubmitHandler.java
@@ -2,17 +2,23 @@ package uk.gov.hmcts.reform.sscs.ccd.presubmit.readytolist;
 
 import static java.util.Objects.requireNonNull;
 
+import java.util.Map;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.sscs.ccd.callback.Callback;
 import uk.gov.hmcts.reform.sscs.ccd.callback.CallbackType;
 import uk.gov.hmcts.reform.sscs.ccd.callback.PreSubmitCallbackResponse;
 import uk.gov.hmcts.reform.sscs.ccd.domain.EventType;
 import uk.gov.hmcts.reform.sscs.ccd.domain.HearingRoute;
+import uk.gov.hmcts.reform.sscs.ccd.domain.RegionalProcessingCenter;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
 import uk.gov.hmcts.reform.sscs.ccd.domain.YesNo;
 import uk.gov.hmcts.reform.sscs.ccd.presubmit.PreSubmitCallbackHandler;
 import uk.gov.hmcts.reform.sscs.helper.SscsHelper;
+import uk.gov.hmcts.reform.sscs.service.RegionalProcessingCenterService;
+import uk.gov.hmcts.reform.sscs.service.hmc.topic.HearingRequestHandler;
+import uk.gov.hmcts.reform.sscs.util.SscsUtil;
 
 @Service
 @Slf4j
@@ -23,13 +29,21 @@ public class ReadyToListAboutToSubmitHandler implements PreSubmitCallbackHandler
     static final String GAPS_CASE_WARNING = "This is a GAPS case, If you do want to proceed, "
             + "then please change the hearing route to List Assist";
 
+    private final RegionalProcessingCenterService regionalProcessingCenterService;
+    private final HearingRequestHandler hearingRequestHandler;
+
+    public ReadyToListAboutToSubmitHandler(@Autowired RegionalProcessingCenterService regionalProcessingCenterService,
+                                       @Autowired HearingRequestHandler hearingRequestHandler) {
+        this.regionalProcessingCenterService = regionalProcessingCenterService;
+        this.hearingRequestHandler = hearingRequestHandler;
+    }
+
     @Override
     public boolean canHandle(CallbackType callbackType, Callback<SscsCaseData> callback) {
         requireNonNull(callback, "callback must not be null");
         requireNonNull(callbackType, "callbacktype must not be null");
 
-        return callbackType.equals(CallbackType.ABOUT_TO_SUBMIT)
-            && callback.getEvent() == EventType.READY_TO_LIST;
+        return callbackType.equals(CallbackType.ABOUT_TO_SUBMIT) && callback.getEvent() == EventType.READY_TO_LIST;
     }
 
     @Override
@@ -40,22 +54,41 @@ public class ReadyToListAboutToSubmitHandler implements PreSubmitCallbackHandler
         }
 
         SscsCaseData sscsCaseData = callback.getCaseDetails().getCaseData();
-        var response = new PreSubmitCallbackResponse<>(callback.getCaseDetails().getCaseData());
 
         if (!sscsCaseData.isIbcCase() && HearingRoute.GAPS == sscsCaseData.getSchedulingAndListingFields().getHearingRoute()) {
+
             if (warningsShouldNotBeIgnored(callback)) {
+                var response = new PreSubmitCallbackResponse<>(callback.getCaseDetails().getCaseData());
                 response.addWarning(GAPS_CASE_WARNING);
                 log.warn("Warning: {}", GAPS_CASE_WARNING);
                 return response;
             }
+
+            return HearingHandler.GAPS.handle(sscsCaseData, hearingRequestHandler);
         }
 
         if (SscsHelper.hasHearingScheduledInTheFuture(sscsCaseData) && warningsShouldNotBeIgnored(callback)) {
+            var response = new PreSubmitCallbackResponse<>(callback.getCaseDetails().getCaseData());
             response.addWarning(EXISTING_HEARING_WARNING);
             log.warn("Warning: {}", EXISTING_HEARING_WARNING);
             return response;
         }
-        return response;
+
+        String region = sscsCaseData.getRegion();
+
+        if (sscsCaseData.isIbcCase()) {
+            SscsUtil.setListAssistRoutes(sscsCaseData);
+            return HearingHandler.valueOf(HearingRoute.LIST_ASSIST.name()).handle(sscsCaseData, hearingRequestHandler);
+        }
+        Map<String, RegionalProcessingCenter> regionalProcessingCenterMap = regionalProcessingCenterService
+                .getRegionalProcessingCenterMap();
+
+        HearingRoute route = regionalProcessingCenterMap.values().stream()
+                .filter(rpc -> rpc.getName().equalsIgnoreCase(region))
+                .map(RegionalProcessingCenter::getHearingRoute)
+                .findFirst().orElse(HearingRoute.GAPS);
+
+        return HearingHandler.valueOf(route.name()).handle(sscsCaseData, hearingRequestHandler);
     }
 
     boolean warningsShouldNotBeIgnored(Callback<SscsCaseData> callback) {
